add delete button for order that are not only in draft

+++ src/application/commands/order/deleteOrderCommand.ts
@@ -8,6 +8,14 @@ export class DeleteOrderCommandHandler {
     constructor(private _orderStore = useOrdersStore()) {}
 
     async handle(cmd: DeleteOrderCommand) {
+        const order = await orderRepository.getById(cmd.id);
+
+        if(!order) return;
+
+        if(order.netAllocatedAmount > 0) {
+            throw Error("Cannot delete an order with an allocated amount larger than zero.");
+        }
+
         await orderRepository.remove(cmd.id);
         this._orderStore.remove(cmd.id);
     }
